Correct 0

+++ 0x02-functions_nested_loops/0-putchar.c
@@ -11,7 +11,7 @@ int main(void)
 {
 	char putch[8] = {'_', 'p', 'u', 't', 'c', 'h', 'a', 'r'};
 
-	int c;
+	long unsigned int c;
 
 
 	for (c = 0; c < sizeof(putch); c++)
